Se ha refactorizado el enum Soporte y se han solucionado bugs relacionados con los contenidos
