+++ src/main/java/es/library/databaseserver/contenido/Contenido.java
@@ -38,6 +38,37 @@ public class Contenido implements AbstractContenido{
 	private Long IDLibro;
 	private Long IDAudiovisual;
 	
+	/**
+	 * Enum que define los soportes en los que pueden estar los contenidos de la biblioteca
+	 * @author Daniel García
+	 *
+	 */
+	public enum Soporte {
+		FISICO(false,false),
+		E_BOOK(false,false),
+		CD(true,true),
+		DVD(true,true),
+		BLURAY(true,true),
+		VINILO(true,false),
+		CASETE(true,false);
+		
+		private boolean multimedia;
+		private boolean audio;
+		
+		Soporte(boolean audio,boolean multimedia){
+			this.multimedia = multimedia;
+			this.audio = audio;
+		}
+		
+		public boolean isMultimedia() {
+			return multimedia;
+		}
+		
+		public boolean isAudio() {
+			return audio;
+		}
+	}
+	
 	public Contenido(Long iD, String titulo, String autor, String descripcion, Integer ano, String idioma,
 			Soporte soporte, boolean prestable, Integer diasDePrestamo, boolean disponible,
 			LocalDate fechaDisponibilidad) {
