+++ src/main/java/es/library/databaseserver/contenido/types/Audio.java
@@ -3,7 +3,6 @@ package es.library.databaseserver.contenido.types;
 import java.time.LocalDate;
 
 import es.library.databaseserver.contenido.Contenido;
-import es.library.databaseserver.contenido.Soporte;
 
 public class Audio extends Contenido{
