+++ src/main/java/es/library/databaseserver/contenido/types/Video.java
@@ -2,8 +2,6 @@ package es.library.databaseserver.contenido.types;
 
 import java.time.LocalDate;
 
-import es.library.databaseserver.contenido.Soporte;
-
 public class Video extends Audio{
 	
 	private Integer edadRecomendada;
